julian/requirements.txt
nltk==3.2.4
Cython==0.26.1
#Keras==2.0.5
#mxnet==0.11.0
#numba==0.35.0+10.g143f70e90
numpy==1.13.1
pandas==0.20.3
#protobuf==3.4.0
scikit-learn==0.19.0
scipy==0.19.1
#seaborn==0.8
#setuptools==36.2.0
#snowballstemmer==1.2.1
#Sphinx==1.6.3
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
#torch==0.1.12.post1
#Flask-RESTful==0.3.6
boto3==1.4.7
#Flask-Cors==3.0.3
msgpack-python==0.4.8
kafka-python==1.3.5
ujson==1.35
